strip test video to 3 frames, only imshow if AUTHOR_TESTING

+++ t/tracking.t
@@ -15,8 +15,7 @@ use PDL::OpenCV::Objdetect;
     my $cpy = $imgb->copy;
     my ($pts) = ellipse2Poly([250,250],[200,100],45,60,120+$_,1);
     rectangle($cpy, $_, $_+1, [255,0,0,0]) for $pts->dog;
-    imshow("ud", $cpy);
-    waitKey(50);
+    imshow("ud", $cpy), waitKey(50) if $ENV{AUTHOR_TESTING};
   }
 }
 
@@ -25,21 +24,19 @@ my $CC_DIR = '';
 my ($loaded) = $cc->load($CC_DIR.'/haarcascades/haarcascade_frontalface_alt.xml') if $CC_DIR;
 die "Failed to load" if $CC_DIR and !$loaded;
 
-my $vfile='t/Megamind.avi';
+my $vfile='t/frames.avi';
 my $vc = PDL::OpenCV::VideoCapture->new;
 die if !$vc->open($vfile, CAP_ANY);
 isnt $vc->getBackendName, undef, 'getBackendName works';
 my ($frame, $res) = $vc->read;
 ok $res, 'read a frame right';
 is_deeply [$frame->dims], [3,720,528], 'right dims' or diag $frame->info;
-my $x = 1;
-($frame, $res) = $vc->read for 1..$x; # blank frames
 
 is my $fcc = PDL::OpenCV::VideoWriter::fourcc(split '', 'MP4V'), 1446269005, 'fourcc right value';
 
-my $box=pdl(qw/169 88 192 257/);
+my $box=pdl(qw/240 161 187 202/);
 my $tr = PDL::OpenCV::TrackerKCF->new;
-if ($box->at(0) == 0) {
+if ($ENV{AUTHOR_TESTING} && $box->at(0) == 0) {
   namedWindow("ud",WINDOW_NORMAL);
   $box = selectROI("ud",$frame,1,0);
   destroyWindow("ud");
@@ -48,6 +45,7 @@ $box = $tr->init(frame_scale($frame),$box);
 
 my $lsd = PDL::OpenCV::LineSegmentDetector->new(LSD_REFINE_STD);
 
+my $x = 0;
 while ($res) {
   ($box, my $track_res) = $tr->update($frame = frame_scale($frame));
   my ($lines) = $lsd->detect(my $gray = cvtColor($frame, COLOR_BGR2GRAY));
@@ -64,9 +62,8 @@ while ($res) {
       rectangle($frame, [$bx,$by], [$bx+$bw,$by+$bh], [0,255,255,0], 2, 1, 0);
     }
   }
-  imshow("ud", $frame);
-  waitKey(1);
-  if ($x<98 || $x > 153 && $x<200) {
+  imshow("ud", $frame), waitKey(500) if $ENV{AUTHOR_TESTING};
+  if ((int($x/2) % 2) == 0) {
           is(all ($box) >0,1,"tracker found box $x.");
           ok $track_res, 'tracker said found';
   } else {
